muudetud punktide kogusumma arvutamise loogikat

+++ 01_07.02_andmemudelid/kymnevoistlus/src/main/java/ee/agnessa/kymnevoistlus/entity/Athlete.java
@@ -21,6 +21,6 @@ public class Athlete {
     private Long id;
     private String name;
     private String country;
-    private Long age;
-    private int totalPoints;
+    private Integer age;
+    private Integer totalPoints;
 }
